requirements.txt
faiss-cpu
numpy
sentence-transformers
ollama
streamlit
pypdf
python-docx